Rename lager package to log
